Updated variant start and end from int to long

+++ variation-commons-core/src/main/java/uk/ac/ebi/eva/commons/core/models/IAnnotation.java
@@ -24,9 +24,9 @@ public interface IAnnotation {
 
     String getChromosome();
 
-    int getStart();
+    long getStart();
 
-    int getEnd();
+    long getEnd();
 
     Set<? extends IXref> getXrefs();
